Add graphs for 1-minute load average like Ganglia
With diamond upgraded (T97635), HOST.cpu.cpu_count is added. The
number of nodes should be calculated by isNonNull() but that is not
available in our version of Graphite. We workaround that by scaling   
the loads to 0, offset them with 1, and finally calculating a sum;    
only non-null values will be scaled and offset-ed, contributing to
the sum.

For some reason the running process counts look larger than that
would have been reported by Ganglia.